shift everything over to the srl reader
