--- doc/api/allennlp.data.dataset_readers.srl_bert_reader.rst
@@ -1,7 +1,0 @@
-allennlp.data.dataset_readers.srl_bert_reader
-=============================================
-
-.. automodule:: allennlp.data.dataset_readers.srl_bert_reader
-   :members:
-   :undoc-members:
-   :show-inheritance:
